# old/tools/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DetravLauncher.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Update/List/{name}": {
      "get": {
        "tags": [
          "Update"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileListModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListModel"
                }
              }
            }
          }
        }
      }
    },
    "/Update/File": {
      "get": {
        "tags": [
          "Update"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Upload/File": {
      "post": {
        "tags": [
          "Upload"
        ],
        "parameters": [
          {
            "name": "apiKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Upload"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apiKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FileListModel": {
        "type": "object",
        "properties": {
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FileModel": {
        "type": "object",
        "properties": {
          "relativePath": {
            "type": "string",
            "nullable": true
          },
          "hash": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "FileUploadModel": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "seek": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      }
    }
  }
}